# docs/swagger.yaml
basePath: /
definitions:
  domain.CategoryNode:
    properties:
      children:
        items:
          $ref: '#/definitions/domain.CategoryNode'
        type: array
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      type:
        type: integer
    type: object
  domain.CategoryStatsItem:
    properties:
      amount:
        type: string
      category_id:
        type: integer
      category_name:
        type: string
      currency:
        type: string
      parent_category_id:
        type: integer
      parent_category_name:
        type: string
    type: object
  domain.CategoryStatsOutput:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.CategoryStatsItem'
        type: array
      month:
        type: string
      type:
        type: integer
    type: object
  domain.DailyStatsDay:
    properties:
      date:
        type: string
      items:
        items:
          $ref: '#/definitions/domain.DailyStatsItem'
        type: array
    type: object
  domain.DailyStatsItem:
    properties:
      currency:
        type: string
      expense:
        type: string
      income:
        type: string
    type: object
  domain.DailyStatsOutput:
    properties:
      days:
        items:
          $ref: '#/definitions/domain.DailyStatsDay'
        type: array
      month:
        type: string
    type: object
  domain.MonthlyStatsItem:
    properties:
      currency:
        type: string
      net_amount:
        type: string
      total_expense:
        type: string
      total_income:
        type: string
    type: object
  domain.MonthlyStatsOutput:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.MonthlyStatsItem'
        type: array
      month:
        type: string
    type: object
...
  handlers.CreateCategoryRequest:
    properties:
      name:
        type: string
      parent_id:
        type: integer
      parent_name:
        type: string
      type:
        type: integer
    type: object
...
  handlers.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Accounting app backend API for auth, categories, transactions, stats,
    and API keys.
  title: Sumi API
  version: 0.1.0
paths:
...
  /api/categories:
    get:
      parameters:
      - default: 1
        description: 'Category type: 1 expense, 2 income'
        in: query
        name: type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.CategoryNode'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List system categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Category payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.CategoryNode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create a second-level category
      tags:
      - Categories
...
  /api/stats/category:
    get:
      parameters:
      - description: Month in YYYY-MM
        in: query
        name: month
        type: string
      - default: 1
        description: 'Transaction type: 1 expense, 2 income'
        in: query
        name: type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CategoryStatsOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get category stats
      tags:
      - Stats
  /api/stats/daily:
    get:
      parameters:
      - description: Month in YYYY-MM
        in: query
        name: month
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.DailyStatsOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get daily stats
      tags:
      - Stats
  /api/stats/home:
    get:
      parameters:
      - description: Month in YYYY-MM
        in: query
        name: month
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MonthlyStatsOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get monthly stats
      tags:
      - Stats
  /api/stats/monthly:
    get:
      parameters:
      - description: Month in YYYY-MM
        in: query
        name: month
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MonthlyStatsOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get monthly stats
      tags:
      - Stats
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Programmatic API key for server-to-server access.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Use the format: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"